modules: openthread: Use OT platform files from Zephyr
In this hybrid approach, we can implement nrf-specific
OpenThread platform files, but all others are sourced
from Zephyr.

For now, there is only the crypto_psa.c file that is specific
for nrf-sdk, but in the future, there will be more of them.

All Kconfigs are sourced from Zephyr, there is additional
Kconfig.features.nrf for defining sdk-nrf related OpenThread
feature Kconfigs.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>